requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
streamlit-extras>=0.3.0
python-dateutil>=2.8.0
pandas>=2.0.0
